requirements.txt
django>=1.11
djangorestframework>=3.7.0
django-debug-toolbar==1.8
restframework-definable-serializer
pillow
